Move the app to /claimstatus
