+++ pages/claimstatus/index.tsx
@@ -7,8 +7,8 @@ import { useRouter } from 'next/router'
 import Link from 'next/link'
 import { GetServerSideProps } from 'next'
 
-import { Header } from '../components/Header'
-import { Footer } from '../components/Footer'
+import { Header } from '../../components/Header'
+import { Footer } from '../../components/Footer'
 
 export default function Home(): ReactElement {
   const { t } = useTranslation('common')
